UT-206: refactor: manually cherry picked footer changes from UT-211/reinplementTabs.
