+++ src/client/components/App.tsx
@@ -7,11 +7,11 @@ import { setActiveTab } from '../../model/reducers/appNavAction'
 // Components:
 import { Thumbnail } from './Thumbnail'
 import { RenderFullHeader } from './Header'
-import { Footer } from './Footer/Footer'
 
 //CSS files:
 import '../css/Rmc-tabs.css'
 import '../css/App.css'
+import { OperationModeFooter } from './Footer/OperationModeFooter'
 
 const channel = new URLSearchParams(window.location.search).get('channel')
 const specificChannel = parseInt(channel) || 0
@@ -30,7 +30,7 @@ export const App = () => {
                         <Thumbnail />
                     </div>
                 </div>
-                <Footer />
+                <OperationModeFooter />
             </div>
         )
     } else {
@@ -46,7 +46,7 @@ export const App = () => {
                         {renderTabData()}
                     </Tabs>
                 </div>
-                <Footer />
+                <OperationModeFooter />
             </div>
         )
     }
