+++ src/client/css/Operation-mode-edit-visibility-footer.css
@@ -1,0 +1,17 @@
+.operation-mode-edit-visibility-footer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.operation-mode-edit-visibility-footer__text {
+    color: #4175e6;
+    font-size: 35px;
+    text-align: center;
+}
+
+.operation-mode-edit-visibility-footer__controls button {
+    margin-left: 10px;
+    background-color: #808080;
+    top: 25%;
+}
